requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
snowflake-connector-python
boto3
fastparquet
azure-storage-blob
pyarrow
numpy
pandas
python-dateutil
pytz
typing-extensions
urllib3
fastparquet
azure.identity
azure.keyvault.secrets